Make depend_on_past look at the task level
TaskInstances now look at the TaskInstance level again
for depend_on_past.

This includes a workaround for DagRuns that have been scheduled
but its TaskInstances not yet. If a previous dagrun is found
but it does not contain the same task_id depend_on_past is
considered on the dag run level.